Add Day 1
Since I haven't seen Kotlin before, I solved the puzzle using arrays,
for loops, moving int "pointers" into arrays etc. The solution works but
I feel that there should probably be a nicer way to do it.

(later...) I've come across the windowed function. It seems to be doing
exactly what I need. I should try to remember it exists if I need it at
some point in future. I'm committing this happy that I managed to
actually write some code in Kotlin, it works and does what it should,
and on top of that it also doesn't look that bad.