Add Day 2
Yesterday the functions were enough but it seems that today the best
approach how to get from part1 to part2 would be to use inheritance and
override just some of the methods.

The program feels weirdly empty. There seem to be too many short or
empty lines for such a few lines of code. I've tried reformatting it in
IDE with no success.

I was happy to find out that descructuring works and that I don't need
to write many lines just to get all the matching groups from a regex.
I'd be happier if there was some nicer way to convert some of those from
Strings to numbers. To be honest, I feel that it must exist, I just
haven't learnt it yet.

It took me a while to understand what "!!" does. I guess I'll need to
take a closer look at now nullable types work.

The other puzzling thing is why Int value cannot be automatically
converted to Long. Some things feel really strange in this language...