Add Day 4
We're playing bingo today. Just when I start thinking about diagonals I
come across the section saying that in the squid bingo, diagonals don't
count. Great, we just saved an extra line of code.

I've managed to learn the difference between init and constructor (but
only after IDE has told me to convert my code to init), found mapIndexed
and successfully applied the knowledge of a windowed function that I
learnt when solving day 1. I've also managed to use the "to" operator
which feels very Kotlin-like (I wonder what would someone who really
knows Kotlin says).

Luckily the input data is constructed the way that there are no
simultaneously winning boards in the same turn (or at least in the first
and the last one) so my code is good enough and the results are OK.