Add Day 5
It seems that I no longer fight the syntax so much. I've reused the
regex parsing and destructuring from day 2 but this time, I've even
managed to get integers out of it without converting them separately.

However, even today came with a surprising moment. In what universe is
x.coerceAtLeast(y) more readable than Math.max(x,y) so that the IDE
would complain that I should replace max with the Kotlin function? I
guess I'll let it for the future readers of this code to decide.