Add Day 6
I open the task and I see a recursion. I'm thinking if this is a good
place to try the Kotlin tailrec approach, but am not so sure about that.
What I am sure about is that part 2 will contain a bigger number of days
so I quickly forget about trying to construct a list of fish in memory.

Luckily, I still remember some tricks from dynamic programming and
caching pre-computed results is something this day calls for. After
writing a function I realize it's Fibonacci, only not f(n-1)+f(n-2) but
1+f(n-7)+f(n-9).

Given the tail is quite long I give up converting it to tailrec function
for the sake of readability. I know that my cache eliminates the
exponential nature of the recursion anyway. To try at least some of
Kotlin features I force myself to re-write my 3-way if to when and I
discover getOrPut funcion of a map. That should be good enough.

After submitting part 1 I learn that I anticipated part 2 correctly, so
I just add one more check to make my "unit tests" complete and
successfully run my code even on the bigger input.