Add Day 7
Today morning I'm in a hurry because I have an appointment to make. I
open the task on mobile to check if I have a chance of making it or I'll
need to delay it to the afternoon.

At first glance it reminds me of a linear regression, only in 1D. Too
bad I don't remember anymore how to do it mathematically. I check the
value ranges and I decide that the range is small enough for the brute
force approach to work. This time I'm lucky that the crabs aren't spread
across the whole long range. I wonder how embarassed I'll be when I come
across some simple solution using math properly.

As it turns out, the only modification needed for part 2 is the fuel
consumption function. In this case I still remember how to calcualte a
sum of arithmetic sequence so I save an extra loop using s simple
formula and successfully submit the results.

In the afternoon I just refactor the utils because I'm repeating same
patterns when parsing the input too many times, I extract the common
crab calculation logic to a separate function to get rid of copy-paste
code and today's solution is ready to be committed.