Add Day 8
After opening the task I immediately spot the segment display. After
reading first lines of the puzzle I suspect it will lead to some digit
parsing or identification.

To my surprise, the first long segment of text just explains which
numbers are uniquely identified by their lenght and since the first task
is just to count those numbers, there's no problem writing it. The fact
that the strings represent some segment display numbers is completely
irrelevant.

Not surprisingly, part two is exactly what I thought it would be. For a
brief moment I consider writing a bruteforce function (7 factorial is
not a big number) but since the author spent so much time describing the
logical elimination process to figure out which number is which, I feel
like I owe it to him to solve the task analytically.

That means that instead of writing code, today's the time for a pencil
and a paper. After some tome I come up with a sequence of steps to
figure out which number is which by comparing the segments.

After typing the sequence into code IDE tells me that some of my
variables representing known segments are unused, so I get rid of them
to realize, that the only standalone segment I needed to identify is
"e".

Once the mapping between the input sequences and actual digits exists,
creating numbers and calculating a sum is trivial.