Add Day 9
Today we're dealing with 2D input so we'll probably have to do some
exploration.

The first part seems to consider only adjacent cells, so I don't need to
write any type of search. On the other hand, it's just day 9 so the
tasks should be relatively easy.

The second part immediately calls for DFS. First I write the the code to
start from the low points and keep marking basins only if the height
increases. It works OK and after some time of coding I earn a gold star.

After a while I realize that if all locations (apart from 9) will always
be part of exactly one basin, it means, that there are no local minima.
I.e. it's not possible for the basin floor to go up and then down again
without reaching 9 first. That would mean that the peak point would
be part of both basins.

With this new observation I realize that if I just cut off the ridges
(height == 9), the whole map splits into components representing
individual basins. Then I just need to find those without checking the
heights inside.

I simplify the algorithm and after verifying that it still yields the
same result I'm committing it to the repository.