Add Day 10
A valid parentheses problem is actually quite common "warm-up task". As
a result of that I spend most of my time exploring standard Kotlin data
structures looking for something that represents a stack. Once I come
across ArrayDeque part 1 is practically solved.

For the second part I only need to add the remaining stack to my
function output. I also decide to change what I store in the stack.
Originally I had the opening brackets there, now I'm storing the closing
ones. It's not completely necessary (it worked even without this
change), but at least this way the score map for the second part
corresponds exactly to the one listed in the task description and does
not contain the opposite brackets.