Add Day 11
Today we're in a simulation mode. The first thing I can think of is the
game of life. However, it turns out that the simulation of the octopus
cavern can be done in-place without a copy of the grid in each step.

I suspect that part two will require me to calculate the flashes after
some ridiculous amount of steps so I start writing a function that takes
the number of steps as an argument and try to design it in a way where
the states can be cached to detect loops.

I submit part one only to learn that my expectations have been
completely incorrect. Once again I learn that the premature optimization
is the root of all evil.

I remove the completely unnecessary parts, move a single step simulation
to a separate method and submit also part 2.

After my stars I earned I decide to beautify my code with some Kotlin
features like type aliases. I immediately realize that the current
design where all days are in the same package is not really ideal as the
aliases pollute the namespace of other days too. I'm not sure why
JetBrains published their AOC template like this. They might have wanted
to keep things simple, however, it doesn't seem to teach the good
practices from the beginning.

I try to move the days to individual packages only to realize that my
paths to input files (which I moved too) now don't match. It would be
probably better if this project had the proper "maven directory
structure", I put the files in src/main/resources and modified the util
function to read them from classpath. After all, then the explicit
source directory declaration in gradle build file wouldn't be needed.

However, that seems like a big refactoring so I revert it back to the
original state and decide to keep it "as is" even with the side effects.

Since the extension functions are added inside main, it should not
affect the other days, so the only visible thing is the type alias.

I also try to remove the magic numbers and introduce named constants
only to learn, that I cannot declare a "const val" inside my main
function and that the IDE complains if I use an UPPERCASE_NAME for my
local val.

It seems that Kotlin and I are still not on the same wavelenght.