Add Day 13
Half-way through the advent I've been reminded (yet again) how important
it is to read the task description properly.

After writing the paper folding logic I spent quite some time wondering
how is it possible that my result does not meet the expectations. Only
later I've realized that I was only supposed to complete the first fold,
not all of them.

However, this way the only thing left for solving the second part was to
write string concatenation logic.

As for Kotlin constructs, I've used the multi-line string for the first
time and I also finally started to see the benefits of a let function.
That means that today was also beneficial from the "learning Kotlin"
perspective.