Add Day 14
Today I've taken a long journey there and back.

First, suspecting that the actual polymer will be massive, I've written
the altorithm correctly just trackint segment counts. All worked until
the point where I needed to count the elements. I realized that if I sum
the elements in the segments I'll get most of them twice (except for the
first and the last one), but instead of adding the first (or last)
element to my polymer data structure and fixing the math, I panicked and
re-written the algorithm to actually construct the polymer.

It worked... for the first part. The second part confirmed my fears and
this time the polymer really was massive, I reverted the algorithm to
the original state, fixed the element count calculation and spend quite
some time being angry at myself for not doing it properly the first
time.

As for Kotlin API, I've spent quite some time searching if there is a
function in a map that would give me the key with the highest/lowest
value (and haven't found any). Only later I realized that it's not the
key I'm after, but the value itself. What is it, that I have been
reminded of yesterday? "Read the task description carefully"? Well...
here it was again.