Add Day 15
Today looks fine. Dijktstra's shortest path algorithm is notoriously
known so hopefully there will be no surprises.

I start by reusing the input parsing from day 9. As expected,
pathfinding implementation is smooth, so I submit part 1 with no issues.

Part 2 seems to be just a slightly bigger grid so I simply replicate it
(it still fits into memory) and apply the same algorithm.

The result is OK and I earn the 2nd star, however, there is a noticable
lag when calculating the result for part 2 (approx. 2s).

I decide that I should probably try to optimize it to run faster.

I suspected my "smart/lazy" grid implementation as a map of points to be
the source of the issue, so I decide to revert to pure int arrays. Since
I don't really like low-level operations to be exposed, I encapsulate
grid implementation and the array-manipulating logic to a SquareGrid
class.

It seems that it really has helped, but I'm still constructing the big
grid explicitly. Since I already have the logic encapsulated, I decide
that I can calculate the risk data on-the-fly when accessing an index
without the need to store all the values in a massive array.

However, I'm facing some issues when trying to anonymously extend my
class and override the method. The compiler isn't really helpful because
it just says "too many arguments for public constructor". It seems that
the syntactic sugar of moving last lambda out of function calls really
has some disadvantages. After considerable time spent Googling I come
across the "object : Class" construct which finally removes the compiler
error.

Now I'm happy with both the result, the code and the execution time so
there's nothing stoping me from committing it to the repository.