Add Day 16
Day 16 is hexadecimal... what a coincidence :)

Again, I'm in a situation where the best time for today is shorter than
it took me to even read and understand the task description.

I see that a naive idea to convert the string to a number and then to
binary digits would need a BigDecimal or something similar, so I decide
to take this as an opportunity to learn how to work with lazy-evaluated
sequences.

I successfully write a (hopefully) lazy evaluated sequence to produce
binary digits from input, however, I learn that the sequence is not
modified when I call take() which complicates my parsing a bit. After
some research I realize that the thing I need is an iterator not a
sequence, but luckily the only required change is in the name of the
function that generates it.

After obtaining the iterator write the recursive parsing function. I
struggle a bit with subpacket size specified in bit length, but adding
sub-iterator solves this part too.

With the structure completely parsed, calculating the results is no
longer complicated.

I just suspect that I'll need to extract the parsing and computation
logic out in the next days, because re-using custom computers running
strange code seems to be the common AoC theme. But at this point I
decide to follow YAGNI and leave it until it's really needed.