Add Day 17
Today looks like another "pen & paper" task.

I start by checking the testing and real input and realize that the
target area is always lower and more to the right than the submarine.
That simplifies things a bit.

I write the function to check if the probe hits the destination for a
given velocity. The observation from above means, that if the probe is
too much to the right or too low, it surely won't hit the target.

Now the question is which velocities to explore.

Since we want to maximixe the probe peak altitude, it doesn't make sense
to shoot down (then the max alt is always 0). The higher we shoot the
bigger the velocity will be when the probe returns to level 0 (and yes,
it is guaranteed that one of the points on the trajectory will be on
level 0). If we want to have a chance of hitting the target in the next
step, the velocity must be lower than the bottom target coordinate.

As for the horizontal part of the velocity, it surely cannot be greater
than the right target coordinate. I could also calculate the minimum
value (i.e. first such N that the sum of 1..N is not lower than left
target coordinate) but given the fact that the coordinates are small I
just decide to count down to 0.

With boundaries in place, I successfully submit the first part.

Part 2 wants me to calculate every velocity. Since I already have some
candidates from part 1, I just replace the search for the maximum with
the sum. However, it turns out that I also need to count the probes that
I shoot downwards. To account for that, I just modify the lower range
bound for "dy" and earn the second star.