Add Day 18
Today was affected by my driving home for Christmas (both the song and
the actual driving). After spending 7 hours in a car I was looking
forward to some nice daily challenge, only to find the snailfish math
homework.

Again, it took me quite a long time to even understand how it should
work. In the end it was not that complicated... at least on paper.

The first struggle with Kotlin was how to define the proper data
structure for the tree where the node can be either a number or another
node. Since I didn't know if Kotlin supports som kind of alternating
types (it probably doesn't) I ended up with a structure containing both
values and left/right nodes (where one of them is empty) and spent quite
some time putting !! to the source code to tell Kotlin that I really
know that a particular value is not null in that situation.

However, after I had all my exclamation marks in place (and even removed
the extra ones I eagerly put there) the code yielded the correct result.

Luckily part 2 was just iteration over pairs with no extra
complications.