Add Day 19
3D transformations are something I cannot do without visual aids. First
I try to come up with the list from the top of my head, but ater four
lines I give up and decide it's time for some "hardware".

I scan the room for something resembling a cube, draw axes on it and
physically rotate it to be sure I got everything.

When all the positions are defined, I just try to rotate and align every
remaining reading to check if at least 12 points match.

The algorithm runs quite a long time so the only thing I can think of is
mobving the known probes coordinates to one set (originally I was
storing rotated scanners individually with all their readings) which
should at least remove duplicate scanning of probes on which the
scanners overlap.

It seems it has helped a bit but the result is still not yielded
instantly but in ca 25s.

However, I'm out of ideas how to optimize it and 25s doesn't seem like
unreasonable time (luckily the input is not much bigger), so it seems
that this is as good as I can make it.