Add Day 20
Today is another game of life variation, only this time, it's infinite.

Simulating the middle part is quite straightforward so the only
remaining thing is the infinite part. However, after realizing that it
will always be 000000000 or 111111111 I just add one extra line to flip
the default value if necessary.

For a brief moment I consider trimming the edges so that my map doesn't
grow infinitely for floating configurations, but the number of steps in
part 2 is not huge and the algorithm runs in reasonable time even
without trimming.