Add Day 22
After reading the decription I quickly come up with a custom Point3D
data class and store the particular points/cubes in a set. That solves
part 1, however, is not good enough for part 2.

I come up with idea to store Range3D (aka cuboid) instead of points and
slice the axis by points mentioned in any instruction. However, even the
first implementation of Set<Range3D> runs long and crashes on
insufficient heap space.

For a while I consider implementing the shape arithmetic (+/-) so that I
do not store more chunks as necessary, but then I decide it would be too
complicated.

I revisit the interval approach but replace the object set with a simple
array (of arrays of bitsets). Luckily this time there are no heap issues
and the algorithm runs as expected.