Add Day 23
Given how restricted the valid moves are and how few amphipods there are
in the list, it took surprisingly long to calculate.

First I wanted to do the calculation purely on the list of amphipods,
however, after it was too long for part2 (I guess list traversal and
filtering really takes long even if there are just 16 elements in the
list) I ended up storing the burrow plan as 2D array and calculating the
paths/destinations explicitly step-by-step.