Add Day 25
The last part of this year's Advent of Code is here. I start reading the
description and I start suspecting that I'll need to find a step where
an empty area of a particular size will open.

Luckily, the task is only to wait until the cucumbers stop moving so a
simple boolean indicator should do the trick.

Since I already own the other 48 stars, there's nothing to do in part 2.

I'm happy that I managed to finish the AoC this year and learned Kotlin
in the process. Some of my predictions (e.g. about reusing the computer
from some of the previous day) have were incorrect, however, I'm not
particularly sad about it. If I could choose, the fewer "assembly
language" tasks the better (even though the one from yesterday turned up
to be quite interesting from programming point of view in the end).