2022 Day 01
Advent of Code is here again. I haven't used Kotlin for a year so it's
time to remember the basics again.

After spending sime time searching for a function to split the list by a
separator I'm making a decision of writing the split logic myself. At
least I'v come across toIntOrNull which takes care of parsing
exceptions.

The rest is just a matter of remembering some of the rich set of
collection transformations.