2022 Day 03
It seems I'm remembering more and more things about Kotlin from last
year.

Intersection of two Sets is already provided by standard functions so
the only thing left is to remember how to split List or String to
chunks.

The first thing that comes to my mind is `windowed`, however, it seems
that the proper way of doing `windowed(x,x)` is `chunked(x)`.

To my surprise `chunked` also exists on String so it can nicely be
applied to part 1 too.