2022 Day 07
Today is about parsing and building the filesystem tree. I come up with
a naive solution witout significant issues, mostly thanks to the fact
that the input is nice (i.e. it doesn't `cd` to same directory multiple
times etc)

After submitting both parts I start optimizing. My first solution
calculated size on demand which is not ideal if you need the size of
each directory, since the nested ones calculate multiple times
(depending how deep they are). I change the calculation to update the
sizes at insertion time.

The second thing I set out to do is to modify the `find` method to
return a `Sequence` instead of `List`. I struggle a bit with how to
reference `this` of encapsulating class (since Java notation of
`FsNode.this` doesn't work in Kotlin). Luckily, I'm saved by Kotlin
Slack that reminds me of `this@FsNode` syntax.

The final piece is the realization that to solve the task I don't need
to keep files in my tree at all and I just need the directories. I
delete the logic related to files and decide to call it a day.