2022 Day 08
I immediately spot the grid which turns out to be the first source of
delay since I cannot decide whether to work with list of Strings or to
parse it to Int array.

For the sake of speed I decide to write initial code without extra
parsing (i.e. on Strings) in imperative code (old habits die hard).

The minor hiccup is the condition in part 2 where I need to stop at the
tree of the same height, however, still include it in the view.

After my solution is OK I start refactoring.

The first thing to do is to convert `List<String>` to `Array<IntArray>`
and re-write of boundary checks to `.indices.contains` (I checked the
source code to be sure it really is implemented in a smart way just by
checking boundaries and does not enumerate over whole range one by one).

Another optimization for better readability (and to practice sequences
from yesterday) is the introduction of the `view` function which yields
a view from a given point in a given direction. This makes `isVisible`
function concise and readable.

I spend some time looking for some sequence operation that would allow
me to stop at the element but still include it, however, after not
finding anything I give up and leave it with the imperative for loop.