2022 Day 09
Today the narrative seems to be very long, luckily it just turs out to
be snake/rope simulation.

I spend some time looking for the correct `sign` function for `Int` and
after finding it the conditions for moving a tail is quite simplified.

Part 2 is just repeating the same logic for 9 chained ropes so no major
modifications are necessary.