2022 Day 10
It's Saturday and I'm afraid that Eric has prepared something more time
consuming today. After opening the task I see `noop` and `addx` which
immediately reminds me of the intcode computer. I expect the worst.

Luckily, there are just 2 commands and one of them does nothing so
simulating the registry value is not that hard.

After unlocking part 2 I read it over and over several times. I come up
with the algorithm but it doesn't yield anything resembling letters. Not
surprisingly, the unit test fails too. However tehe first line seems to
be correct and after examining the register values I relize that when
drawing the 2nd line the value to check is not 40 but it starts from 0
again. This fixes the unit test and the output now renders letters too.