2022 Day 12
Today it's time for BFS. As usually, I forget adding the exploration
candidates to visited list (i.e. I only add them to visited when I
actually take them out of candidate list and do the actual exploration)
which, not surprisingly, leads to super slow algorithm that keeps adding
the same stuff over and over again to the candidate list.

When I now think about it, making it a linked set would work too,
however marking them as visited at the time of adding them as candidates
works OK too.

Part 2 is just the same thing in the opposite direction, just with a
minor twist that the set of potential ends is bigger than one, so
refactoring the code to support it does not come up with any massive
complications.