2022 Day 13
Yesterday I've told someone in Kotlin Slack that on some days input
parsing takes more effort than actually solving the task.

This is one of those days so I spend most of my time writing a parser
for nested `Int` lists.

Then I study the documentation how to override/add comparison operators
for classes and I add one to the `List`.

After submitting part 2 I realize that `sort` is too complex for what we
need to achieve so I refactor the solution to a linear one.