2022 Day 14
As the task description suggests, the task of filling 2D space with
falling material is something I've come across before.

I use the hashmap indexed by points as a cave representation and run the
simulation.

Part 2 requires just a minor modification in both the condition where
the individual grain stops and the one determining when to stop adding
more sand.