2022 Day 16
It's getting harder and it's starting to interfere with my school run
schedule :( However, the break proved out to be beneficial since I've
realized I can make bigger steps instead of just 1 room at a time.

Finding paths to open valves (thanks to wave algorithm) and doing just
these long jumps turned out to be the the piece I was missing.

Part 2, however, really messed up my approach, since I'd need to stop
mid-step as times in desired valve rooms are not aligned for 2 actors.
Luckily there don't seem to be that many valves to bruteforce (i.e.
splitting responsibilities for individual valves) yields a result in a
2-3 mins, which is very likely far from optimal, but still good enough
to earn the 2nd star.