2022 Day 17
Driving home for Christmas... top to toe in tailbacks...

I've spent 8 hours driving in not really pleasant conditions, as it was
snowing for the last few hours of the journey. I wasn't sure I'd find
power to do the task today, but the Elves really needed me.

The straitforward (though lot of writing) Tetris simulation did the
trick for the first part, but part 2 was too big to simulate.

I've remembered I've encountered similar task before where the game of
life (or some pots in a cavern?) repeated so I decided to write an
algorithm to check for loops. After struggling with how to represent the
state using data classes of `(Int,Int,Set)` I ended up with simple
`String` representing the state as it was fastest to work with.

The loop has been found quite soon (and quite short) and it earned me
the second start (fruit). I'm starting to have a feeling that the
reindeer might have a chance of getting some star fruit full of magical
energy even if it turns out that something has happened to the grove in
the end.