2022 Day 19
After Saturday I was thinking that the hard part was over and that the
tasks scheduled for the work week would not be that hard. I couldn't be
more wrong.

Today I spent most of the time thinking about how to reduce the
traversed space so that it finishes in a reasonable time with reasonable
memory. The first optimization, i.e. to conunt the future potential,
only helped a bit and even though the space has shrunk, it was still
failing on insufficient heap space for 2nd blueprint in part 2.

Luckily, since I'm home, my brother came up with a clever idea, that
since we can only produce one robot at a time, it doesn't make sense to
make more robots of a particular kind than what's the maximum required
by any recipe, because it will always keep replenishing.

This reduced the search space quite significantly, and even though it
still runs tens of seconds, at least it now finished under a minute and
did not complain about insufficient heap.