2022 Day 20
Mixing requires a special type of list that allows quick
insertion/removal at a specified position but also quick direct access
to nth element (or at least the possibility to obtain such reference in
advance).

After some time trying to search the stdlib and trying to be smart about
calculating where to insert and where to remove, I decide to write own
circular list implementation.

The twist in part 2 is just highlights the "strong" type system of
Kotlin. How is it possible that `Long % Int` is `Long` if the remainder
clearly must be smaller than modulus, so if modulus is `Int` the result
should also be `Int`? Luckily, extra `.toInt()` solves the compiler
issues and my code earns me also 2nd star.