2022 Day 21
Parsing expression tree for part 1 contains no surprises apart from the
fact that the result doesn't fit to `Int`.

Part 2, however, requires more logic. I try the naive approach, i.e. to
just construct and print the expression and let Wolfram Alpha solve it.
Unfortunately, the expression is too long so I need to simplify it.

Luckily it turns out that the expression is simple enough so that each
monkey is referenced at most once (i.e. the tree branches don't join)
and that the expressions never divide with "humn" value. As a result of
that the whole expression can be simplified to `k*humn + q = r` where k,
q and r are fractions (i.e. rational numbers).

I spend considerable amount of time writing `Rational` class that is
capable of the required four operations. After that is done, the result
is just a matter of doing `(r-q)/k`.