2022 Day 25
It is the end of the journey. Even though it took a while to realize how
to convert number to snafu, it turns out to be quite simple in the end.
Of course, the first result is wrong, as the result does not fit into
`Int` but (as many times before), `Long` does the trick.