Refactor test input retrieval
There's no need for `getTestInput` method to accept the classifier if
the only thing it does is pass it down to `getInput`. In such case the
calls to `getInput(String)` and `getTestInput(String)` are the same and
for consistency purposes it's better if all such calls use just one way
of retrieving the particular input.

Use kotlin.io.Path instead of java.io.File for file manuipulation
(imported from recent changes in official Kotlin template).