2023 Day 4
What a poor elf. Before he met me, he was full of joy and expectations,
trying to figure out what he's won. After we've figured it out together,
it seems like the only thing that a scratchcard is good for is winning
virtual copies of another scratchcards (which can win more copies of
virtual scratchcards). Recursive lotteries are the worst.