2023 Day 13
Ouch, that hurt, even if nobody was watching. I just wonder, how was I
able to do a closer inspection to find out that each mirror has exactly
one smudge if the point of this whole calculation was to figure out
where the mirrors are. And if I already knew where the mirrors were (and
knew that each of them had exactly one smudge), why did I need to
calculate their positions again? :thinking: