2023 Day 15
The Lava Production Facility is hopefully finally online and the lava
can flow. But I thought the same thing yesterday after aligning the
reflector dish. So can the lava really flow now? That remains to be
seen, however, something tells me that there will be some obstacles on
its way.