2023 Day 20
The parts were in place, just the cables were connected in a "unique"
way. What a poor elf that would need to press a button mor than 2e14
times. Now at least the sand is falling to Island Island, but it would
need to be processed first before it can be used for water filtering.

On a more positive note, I'm already more than half-way on our journey
back, so I am optimistic about being able to fix the snow machines just
in time for Christmas (at least for that part of world that celebrates
on 25th, not on Christmas Eve).