2023 Day 21
Whew... fortunately we were just calculating the steps out of boredom
until there was enough sand. If this was some critical operation, we
might have missed the deadline. Luckily the gardener has designed his
garden in a nice way so that he can easily calculate the number of plots
he can reach in a given (again a convenient one) number of steps.