2023 Day 22
I'm lucky that the sand bricks are stable in horizontal direction and
even if the brick is supported only on one end/corner, there's no force
that would cause it to topple over and fall. If that was the case, I'd
be much more afraid of disintegrating those calculated bricks.