2023 Day 24
The Christmas is doomed. The hailstorm is too much for a progammer
helper to be able to handle on a Christmas day. If the elves need help,
they should try to solve the complicated tasks earlier. In the end I've
managed to break the hailstones, unfortunately it has not been in time
for the snow machine to be saved for Christmas.